requirements.txt
tweepy
openai
python-dotenv
fastapi
uvicorn
httpx
supabase
